pci: Separate device and communication code
PCI device handling code in `device` module; server code in `main`.
