+++ pci/src/main.rs
@@ -10,221 +10,21 @@
 #![no_std]
 #![no_main]
 
-use core::arch::asm;
 use core::str;
 
 use euralios_std::{println, syscalls, message::pci,
                    message::{self, Message},
                    server::{FileLike, DirLike, handle_directory},
                    syscalls::STDIN};
-use core::fmt;
 
 extern crate alloc;
 use alloc::collections::btree_map::BTreeMap;
 use alloc::{string::String, sync::Arc};
 use alloc::format;
-
 use spin::RwLock;
 
-const CONFIG_ADDRESS: u16 = 0xCF8;
-const CONFIG_DATA: u16 = 0xCFC;
-
-#[derive(Clone, Copy)]
-struct PciLocation {
-    bus:  u16,
-    slot: u16,
-    function: u16
-}
-
-impl fmt::Display for PciLocation {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-           write!(f, "{:04X}:{:02X}:{:02X}",
-                  self.bus, self.slot, self.function)
-    }
-}
-
-impl PciLocation {
-    fn from_address(address: u32) -> PciLocation {
-        PciLocation{
-            function:((address >> 8) & 0b111) as u16,
-            slot:((address >> 11) & 0b1_1111) as u16,
-            bus:((address >> 16) & 0xFF) as u16
-        }
-    }
-
-    /// Return PCI bus address
-    fn address(&self) -> u32 {
-        0x8000_0000
-            | ((self.bus  as u32) << 16)
-            | ((self.slot as u32) << 11)
-            | ((self.function as u32) <<  8)
-    }
-
-    fn read_register(&self, register: u8) -> u32 {
-        let addr = self.address()
-            | ((register as u32) << 2);
-
-        let value: u32;
-        unsafe {
-            asm!("out dx, eax",
-                 in("dx") CONFIG_ADDRESS,
-                 in("eax") addr,
-                 options(nomem, nostack));
-
-            asm!("in eax, dx",
-                 in("dx") CONFIG_DATA,
-                 lateout("eax") value,
-                 options(nomem, nostack));
-        }
-        value
-    }
-
-    fn write_register(&self, register: u8, value: u32) {
-        let addr = self.address()
-            | ((register as u32) << 2);
-
-        unsafe {
-            asm!("out dx, eax",
-                 in("dx") CONFIG_ADDRESS,
-                 in("eax") addr,
-                 options(nomem, nostack));
-
-            asm!("out dx, eax",
-                 in("dx") CONFIG_DATA,
-                 in("eax") value,
-                 options(nomem, nostack));
-        }
-    }
-
-    /// Return the Device which is at this PCI bus location
-    /// May return None if there is no device
-    fn get_device(&self) -> Option<Device> {
-        let reg_0 = self.read_register(0);
-        if reg_0 == 0xFFFF_FFFF {
-            return None // No device
-        }
-
-        let vendor_id = (reg_0 & 0xFFFF) as u16;
-        let device_id = (reg_0 >> 16) as u16;
-
-        let reg_2 = self.read_register(2);
-
-        let revision_id = (reg_2 & 0xFF) as u8;
-        let prog_if = ((reg_2 >> 8) & 0xFF) as u8;
-        let subclass = ((reg_2 >> 16) & 0xFF) as u8;
-        let class = ((reg_2 >> 24) & 0xFF) as u8;
-
-        let reg_3 = self.read_register(3);
-
-        let header_type = ((reg_3 >> 16) & 0xFF) as u8;
-
-        let subsystem_id = if header_type == 0 {
-            let reg_B = self.read_register(0xB);
-            ((reg_B >> 16) & 0xFFFF) as u16
-        } else { 0 };
-
-        Some(Device {
-            location: self.clone(),
-            vendor_id,
-            device_id,
-            class,
-            subclass,
-            prog_if,
-            revision_id,
-            header_type,
-            subsystem_id
-        })
-    }
-}
-
-/// Information about a PCI device
-struct Device {
-    location: PciLocation,
-    vendor_id: u16, // Identifies the manufacturer of the device
-    device_id: u16, // Identifies the particular device. Valid IDs are allocated by the vendor
-    class: u8, // The type of function the device performs
-    subclass: u8, // The specific function the device performs
-    prog_if: u8, // register-level programming interface, if any
-    revision_id: u8, // revision identifier. Valid IDs are allocated by the vendor
-    header_type: u8,
-    subsystem_id: u16
-}
-
-impl Device {
-    fn class_str(&self) -> &'static str {
-        match self.class {
-            0x0 => match self.subclass {
-                0 => "Non-VGA-Compatible Unclassified Device",
-                1 => "VGA-Compatible Unclassified Device",
-                _ => "Unknown",
-            },
-            0x1 => match self.subclass {
-                0x0 => "SCSI Bus Controller",
-                0x1 => "IDE Controller",
-                0x2 => "Floppy Disk Controller",
-                0x3 => "IPI Bus Controller",
-                0x4 => "RAID Controller",
-                0x5 => "ATA Controller",
-                0x6 => "Serial ATA Controller",
-                0x7 => "Serial Attached SCSI Controller",
-                0x8 => "Non-Volatile Memory Controller",
-                _ => "Mass Storage Controller"
-            }
-            0x2 => match self.subclass {
-                0x0 => "Ethernet Controller",
-                0x1 => "Token Ring Controller",
-                0x2 => "FDDI Controller",
-                0x3 => "ATM Controller",
-                0x4 => "ISDN Controller",
-                0x5 => "WorldFip Controller",
-                0x6 => "PICMG 2.14 Multi Computing Controller",
-                0x7 => "Infiniband Controller",
-                0x8 => "Fabric Controller",
-                _ => "Network Controller"
-            }
-            0x3 => match self.subclass {
-                0x0 => "VGA Compatible Controller",
-                0x1 => "XGA Controller",
-                0x2 => "3D Controller (Not VGA-Compatible)",
-                _ => "Display Controller"
-            }
-            0x4 => match self.subclass {
-                0x0 => "Multimedia Video Controller",
-                0x1 => "Multimedia Audio Controller",
-                0x2 => "Computer Telephony Device",
-                0x3 => "Audio Device",
-                _ => "Multimedia Controller"
-            }
-            0x5 => match self.subclass {
-                0x0 => "RAM Controller",
-                0x1 => "Flash Controller",
-                _ => "Memory Controller"
-            }
-            0x6 => match self.subclass {
-                0x0 => "Host Bridge",
-                0x1 => "ISA Bridge",
-                0x2 => "EISA Bridge",
-                0x3 => "MCA Bridge",
-                0x4 => "PCI-to-PCI Bridge",
-                0x5 => "PCMCIA Bridge",
-                0x6 => "NuBus Bridge",
-                0x7 => "CardBus Bridge",
-                0x8 => "RACEway Bridge",
-                0x9 => "PCI-to-PCI Bridge",
-                0xA => "InfiniBand-to-PCI Host Bridge",
-                _ => "Bridge"
-            }
-            _ => "Unknown"
-        }
-    }
-}
-
-impl fmt::Display for Device {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{} [{:04X}:{:04X}] {}",
-               self.location, self.vendor_id, self.device_id, self.class_str())
-    }
-}
+mod device;
+use device::{PciLocation, Device};
 
 impl DirLike for Device {
     fn get_dir(&self, _name: &str) -> Result<Arc<RwLock<dyn DirLike + Send + Sync>>, syscalls::SyscallError> {
